remove email component
