+++ src/app/tripview/invite-members-component/friend-list/friend-item.component.css
