+++ src/app/tripview/invite-members-component/friend-list/friend-item.component.html
@@ -2,5 +2,5 @@
 		<img class="circle-sm"  
 		[src]="sanitize('data:image/jpg;base64,' + (_Inviteduser | async)?.profileImage)">
     <p class="friend-name">{{friend.name}}</p>
-		<input type="checkbox" [disabled]="(_hasInvitedUser2)&&(_hasInvitedUser1)" [checked]="(_hasInvitedUser2)&&(_hasInvitedUser1)"/>
+		<input type="checkbox" [disabled]="(_hasInvitedUser2)||(_hasInvitedUser1)" [checked]="(_hasInvitedUser2)||(_hasInvitedUser1)"/>
 </div>
