+++ src/app/tripview/invite-members-component/friend-list/friend-list.component.css
@@ -20,3 +20,4 @@
 	margin-left: -40px;
 	list-style-type: none;
 }
+
