+++ src/app/tripview/invite-members-component/friend-list/friend-list.component.html
@@ -1,0 +1,15 @@
+
+					
+						
+
+
+<div class="att-list">
+	<ul>
+		<li *ngFor="let friend of friends">
+		<friend-item [friend]="friend" [tripId]="tripId"></friend-item>
+		</li>
+	</ul>
+</div>						
+						
+					
+
