--- src/app/tripview/invite-members-component/invite-byemail/invite-byemail.component.html
@@ -1,28 +1,0 @@
-	<div class="invitationForm" >
-			<form  #invitForm >
-				
-				<div *ngIf="_error_flag">
-					<div class="help-block"> {{ _error }} </div>
-				</div>
-				<div class="form-group">
-					<!-- use label for to allow clicking on label to navigate cursor to the specific input field -->
-					<input 
-					type="email"
-					id="email" 
-					class="form-control"				 
-					ngModel
-					name="email"
-					required
-					pattern="^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$"
-					#email="ngModel"
-					placeholder="email">
-					<span class="help-block" *ngIf="(!email.valid && email.touched)">Please enter a valid email!</span>
-				</div>
-				<div class="form-group">
-					<button 
-						class="invite" 
-						[disabled]="!email.valid" 
-						(click)="inviteMembers(email); false">Invite</button>
-				</div>
-			</form>
-	</div>
