--- src/app/tripview/invite-members-component/invite-fb/invite-fb.component.html
@@ -1,8 +1,0 @@
-
-<div class="att-list">
-  <ul>
-    <li *ngFor="let friend of friends">
-      <friends-list [friend]="friend" [tripId]="tripId"></friends-list>
-    </li>
-  </ul>
-</div>
