+++ src/app/tripview/invite-members-component/invite-members-component.component.html
@@ -2,18 +2,39 @@
   <div class="splash-inner splash-open"> 
 		<div class="splash-content text-center">
 			<h3 class="title">Invite Members</h3>
-			<div id="leftnav">
-				<ul>
-					<li><div [routerLinkActive]="['active']" class="link" [routerLink]="['invite_by_email']">Email</div></li>
-					<li><div [routerLinkActive]="['active']" class="link" [routerLink]="['invite_by_fb']">Facebook</div></li>				
-				</ul>
-			</div>
 		</div>
-  		<div class="splash-content2 text-center">
-			<div id="choices">
-				<router-outlet></router-outlet>			
-			</div>
-			<button class="btn btn-lg btn-outline" (click)="closeInvitation()">Cancel</button>
+		<div class="splash-content2 text-center">
+			<div class="invitationForm" >
+				<form  #invitForm >
+					<div class="form-group">
+						<div *ngIf="_error_flag">
+							<div class="help-block"> {{ _error }} </div>
+						</div>
+						<!-- use label for to allow clicking on label to navigate cursor to the specific input field -->
+						<input 
+						type="email"
+						id="email" 
+						class="form-control"				 
+						ngModel
+						name="email"
+						required
+						pattern="^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$"
+						#input="ngModel"
+						placeholder="email or facebook friend's name"
+						(keyup)=search(input.value)
+						>
+						<span class="help-block" *ngIf="(!input.valid && !matchname && input.touched)">Please enter a valid email or a name!</span>
+					</div>			
+					<friend-list [friends]="filtered" [tripId]="tripId"></friend-list>
+					<div class="form-group">
+									<button 
+										class="invite" 
+										[disabled]="!input.valid" 
+										(click)="inviteMembers(input); false">Invite</button>
+					</div>
+				</form>						
+					<button class="btn btn-lg btn-outline" (click)="closeInvitation()">Cancel</button>
+				</div>
 		</div>
 	</div>
 </div>
